Front End Styling and refactoring

+++ frontend/src/Components/Task.js
@@ -1,30 +1,35 @@
-import React, { useState } from 'react'
-import { useNavigate } from 'react-router-dom'
+import React, { useState } from "react"
+import { useNavigate } from "react-router-dom"
 
-import { getTask } from '../data/getTask'
-import { deleteTask } from '../data/deleteTask'
-import './Task.css'
+import { getTask } from "../data/getTask"
+import { deleteTask } from "../data/deleteTask"
+import "./Task.css"
 
 function Task({ task }) {
   const navigate = useNavigate()
   function handleDetails() {
-    console.log('details button clicked!')
+    console.log("details button clicked!")
     getTask(task.taskid).then((response) => {
       console.log(response.data)
-      navigate('/TaskDetail', { state: { taskInfo: response.data } })
+      navigate("/TaskDetail", { state: { taskInfo: response.data } })
     })
   }
   function handleDelete() {
-    console.log('delete button clicked!')
-    deleteTask({
-      id: localStorage.getItem('userid'),
-      taskId: task.taskid,
-    }).then((response) => console.log(response))
+    console.log("delete button clicked!")
+    const shouldDelete = window.confirm(
+      `Are you sure you want to delete the task: ${task.title}?`
+    )
+    if (shouldDelete) {
+      deleteTask({
+        id: localStorage.getItem("userid"),
+        taskId: task.taskid,
+      }).then((response) => console.log(response))
+    }
   }
   return (
-    <>
+    <div className="taskContainer">
       <div className="taskDiv">
-        <h2>{task.title}</h2>
+        <h3>{task.title}</h3>
         <div className="buttonDiv">
           <button className="detailsButton" onClick={handleDetails}>
             Details
@@ -34,7 +39,7 @@ function Task({ task }) {
           </button>
         </div>
       </div>
-    </>
+    </div>
   )
 }
